+++ includes/class-block-editor.php
@@ -102,6 +102,7 @@ class BitStream_Block_Editor {
             'like_nonce' => wp_create_nonce('bitstream_like_nonce'),
             'load_more_nonce' => wp_create_nonce('bitstream_load_more_nonce'),
             'og_fetch_nonce' => wp_create_nonce('bitstream_og_fetch_nonce'),
+            'media_upload_nonce' => wp_create_nonce('bitstream_media_upload_nonce'),
             'post_id' => $current_post_id,
             'poster_url' => class_exists('BitStream_Shortcodes') ? BitStream_Shortcodes::get_poster_page_url() : home_url('/bitstream/')
         ]);